requirements.txt
flask
gunicorn
Flask-SQLAlchemy
flask_session
google-generativeai